requirements.txt
gradio
langchain_ollama 
langgraph 
langchain_google_genai
gradio 
langchain_community 
pypdf
huggingface_hub
pydantic==2.10.6
python-dotenv
langgraph-checkpoint-postgres
langchain_postgres